initial import of meeting output
The present code was entirely written during a microhackathon at HHU Düsseldorf
in late summer 2023; this commit is basically the output of that wrapped in a
package and documented. Except for me, notable contributors include Mo Wilken,
his 2 bicycles, Schweinhaxen, Uerige Altbier, rumors about autocomplete updates
in Julia 1.10, and the guy who couldn't take less than 20 bucks credit card
payment so we had to eat a nice extra plum cake with like 1 liter of whipped
cream each, which was indeed a welcome sacrifice to the altar of science.